switch to node-pre-gyp
