+++ .gitignore
@@ -12,3 +12,8 @@ demo/train-*
 coverage/
 .DS_Store
 tensorflow-tfjs-node*.tgz
+lib/
+build-tmp-napi-v3/
+build-tmp-napi-v4/
+*.tar.gz
+*.zip
