+++ .npmignore
@@ -1,7 +1,6 @@
 .vscode/
 src/**/*_test.ts
 demo/
-deps/tensorflow/*
 scripts/build-npm.sh
 scripts/make-version
 scripts/tag-version
@@ -19,3 +18,9 @@ yarn.lock
 yalc.lock
 yarn-error.log
 .yalc/
+lib/
+deps/
+build-tmp-napi-v3/
+build-tmp-napi-v4/
+*.tar.gz
+*.zip
