+++ package.json
@@ -13,12 +13,14 @@
   },
   "scripts": {
     "build": "tsc",
+    "build-binary": "./scripts/build-and-upload-binary.sh",
     "build-npm": "./scripts/build-npm.sh",
     "build-npm-gpu": "./scripts/build-npm-gpu.sh",
     "clean-deps": "rm -rf deps",
     "coverage": "nyc ts-node src/run_tests.ts",
-    "enable-gpu": "node scripts/install.js gpu download && yarn",
+    "enable-gpu": "node scripts/install-from-source.js gpu download && yarn",
     "format": "clang-format -i -style=Google binding/*.cc binding/*.h",
+    "install-from-source": "node scripts/install-from-source.js",
     "install": "node scripts/install.js",
     "link-local": "yalc link",
     "lint": "tslint -p . -t verbose",
@@ -49,9 +51,22 @@
     "@tensorflow/tfjs": "~1.2.1",
     "adm-zip": "^0.4.11",
     "bindings": "~1.3.0",
+    "edit-json-file": "^1.2.1",
     "https-proxy-agent": "^2.2.1",
     "progress": "^2.0.0",
     "rimraf": "^2.6.2",
-    "tar": "^4.4.6"
+    "tar": "^4.4.6",
+    "node-pre-gyp": "0.13.0"
+  },
+  "binary": {
+    "module_name": "tfjs_binding",
+    "module_path": "./lib/binding/napi-v{napi_build_version}",
+    "host": "https://storage.googleapis.com/tf-builds/pre-built-binary",
+    "remote_path": "./napi-v{napi_build_version}/{version}/",
+    "napi_versions": [
+      3,
+      4
+    ],
+    "package_name": "CPU-linux-1.2.1.tar.gz"
   }
 }
